Support non-`String` values in `fromValue()` methods of generated `enum` classes (#235)
Fixed 2 bugs related to the `fromValue()`-method in generated `enum` classes. The first, affected **all** generated `enum` classes where the `type`-property was not set to `string`. The `fromValue()`-method would always generate the method parameter as `String`, leading to issues caused by comparing incompatible types. The second issue is encountered under the same conditions, but when the configOption `useEnumCaseInsensitive` is set to `true`. This would invoke `equalsIgnoreCase` on non-`String` objects - which caused compilation errors. Both of these bug fixes are contained to the `fromValue()`-method and are fully backwards-compatible.