Support OpenAPI `x-class-extra-annotation` property  (#209)
Adds support for the OpenAPI property `x-class-extra-annotation` (vendor-extension). The property can be set either as a string (for one annotation), or as a literal block `|-` (for multiple annotations). The property is only applicable for `record` classes. The property _could_ conflict with `additionalModelTypeAnnotations` when referring to the same annotation (_if the annotation is not **repeatable**_).