Update Project Version in `.mustache`-files via Maven (#262)
Automatically updates the version numbers in all `.mustache`-files via `maven-resources-plugin`. This is to reduce risks and improve maintainability when doing future version updates. This is for internal use only. It has been verified that the packaged `.mustache`-files are unchanged and have not been relocated due to this change. These `.mustache`-files are now tracked via git.