Support `enumUnknownDefaultCase`-ConfigOption (#365)
Adds support for the `openapi-generator-maven-plugin` ConfigOption
`enumUnknownDefaultCase`, which modifies generated `enum` classes. When
set to `true`, generated `enum` classes will have an additional enum
constant `UNKNOWN_DEFAULT_OPEN_API`. This `enum` constant will be used
as a _default_ return value from the `fromValue(T value)`-method when
none of the enum constants match the provided `value`. This
`configOption` is completely optional, and is set to `false` by default.
This `configOption` is compatible with `useEnumCaseInsensitive`.