Support Generating `enum` classes with `URI` format (#379)
Adds support for generating `enum` classes with `type: string` and `format: uri`. This supports both stand-alone `enum` classes and inner `enum` classes. This was previously not supported, due to two issues. The first issue was due to not importing the `URI` class. The second was an issue inherited from `openapi-generator-maven-plugin`, which causes compilation issues when the `useEnumCaseInsensitive` is set to `true`. Both of these issues have been resolved as part of this change.