Remove `serialVersionUID`-field from Generated `record` Classes (#488)
Removes the `serialVersionUID`-field from generated `record` classes, when the ConfigOption `serializableModel` is set to `true`. The field was removed due to overriding the default value, _`1L` instead of `0L`_, which could cause serialization issues, or was just redundant. As `openapi-generator-maven-plugin` provides no way to set the `serialVersionUID`, the Java default value (`0L`) is a more appropriate value than the previous `1L`.