[webclient] Annotate `record` Constructor with `@JsonCreator` (#491)
Annotates the canonical `record` Constructor with `@JsonCreator`, when the `serializationLibrary` is `jackson` (_default for the `webclient`-library_). This is an expected **breaking** bug-fix, as the _canonical_ constructor will now be used when deserializing, instead of the _compact_ constructor. This _should_ allow fields to be assigned _default_ values, done by the _canonical_ constructor.