Support `x-extra-field-annotation`-property as YAML List/Array (#523)
Fix for a bug in the `pojo.mustache`-file, inherited from `openapi-generator-maven-plugin`, which would not annotate fields correctly when setting `x-extra-field-annotation` using `[]` or `-`. This issue was fixed in the `v7.17.0` release, but has to be applied to the custom mustache templates in this project as well. **NOTE:** this is a bug fix for the OpenAPI-property `x extra-field-annotation` and is a non-breaking bug fix.